pyproject.toml
[project]
name = "miniclue-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "srleom",email = "shengrui.leom@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "boto3 (>=1.38.46,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "app" }
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.poetry.scripts]
start = "app.runner:start"

